[core] move util/gl_helper.hpp to gl/gl_helper.hpp
